requirements.txt
faster-whisper>=1.0.0
openai>=1.0.0
pyyaml>=6.0
httpx>=0.25.0
